feat: 定时更新群信息

+++ plugin/remote_plugin.go
@@ -3,6 +3,7 @@ package plugin
 import (
 	"bytes"
 	"crypto/md5"
+	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -148,6 +149,8 @@ func (p *RemotePlugin) Handle(_ *gorm.DB, ctx *openwechat.MessageContext) (bool,
 	filename := response.Filename
 	msgType, _ := response.Type.Int64()
 	switch int(msgType) {
+	case -1:
+		return false, nil
 	case 1:
 		_, _ = ctx.ReplyText(response.Body)
 	case 2:
@@ -191,23 +194,35 @@ func (p *RemotePlugin) Handle(_ *gorm.DB, ctx *openwechat.MessageContext) (bool,
 }
 
 func (p *RemotePlugin) download(filename string, src string) (io.ReadCloser, error) {
-	resource, err := p.client.R().Get(src)
-	if err != nil {
-		return nil, err
+	if strings.HasPrefix("BASE64:", src) {
+		srcBytes, err := base64.RawStdEncoding.DecodeString(strings.TrimPrefix(src, "BASE64:"))
+		if err != nil {
+			return nil, errors.New("解析资源信息出错")
+		}
+		err = os.WriteFile(filename, srcBytes, 0644)
+		if err != nil {
+			return nil, errors.New("获取资源信息出错")
+		}
+		return os.Open(filename)
+	} else {
+		resource, err := p.client.R().Get(src)
+		if err != nil {
+			return nil, err
+		}
+		body := resource.Body()
+		// 缓存
+		out, err := os.Create(filename)
+		if err != nil {
+			return io.NopCloser(bytes.NewReader(body)), nil
+		}
+		_, err = out.Write(body)
+		if err != nil {
+			_ = out.Close()
+			_ = os.Remove(filename)
+			return nil, errors.New("获取资源信息出错")
+		}
+		return os.Open(filename)
 	}
-	body := resource.Body()
-	// 缓存
-	out, err := os.Create(filename)
-	if err != nil {
-		return io.NopCloser(bytes.NewReader(body)), nil
-	}
-	_, err = out.Write(body)
-	if err != nil {
-		_ = out.Close()
-		_ = os.Remove(filename)
-		return nil, errors.New("获取资源信息出错")
-	}
-	return os.Open(filename)
 }
 
 type (
@@ -228,7 +243,7 @@ type (
 	}
 	remotePluginResponse struct {
 		Error    string      `json:"error"`    // 错误信息，空表示没错误
-		Type     json.Number `json:"type"`     // 回复类型 0:不回复,1:文本,2:图片,3:视频,4:文件
+		Type     json.Number `json:"type"`     // 回复类型 -1:不处理,0:不回复,1:文本,2:图片,3:视频,4:文件
 		Body     string      `json:"body"`     // 回复内容,type=1时为文本内容,type=2/3/4时为资源地址
 		Filename string      `json:"filename"` // 文件名称
 	}
